refactor: `get_tasks` to use tasksDatabase instead of the current days
This change requires that TASKS_DATABASE_URL is changed. It needs to be
of the master task database with a view that is the same as the current
day's linked database of task.

This will result in a speed boost of appoximatly 3~ seconds on the
`get_tasks` script (8.5s -> 5.5s). We get this saving as we don't have
to navigate into the current day's page anymore.