feat: further optimizing the notion-py call by monkey-patching out cache
The cache for notion-py is used to verify that nothing has changes since
the last snapshot since it collected data. Loading and saving the cache
took up some time in the startup/teardown of a call. The larger the
cache the longer it took to load back in.

In my case, everything is happening really quickly as short script
executions. I decided to monkey-patch out the saving/loading of the
cache for massive speed improvments.

With caching:
  No cache present:
    client() = 0.07s
    get_tasks.py = 1.53s

  Cache present:
    client() = 0.41s
    get_tasks.py = 2.08s

Without caching:
  client() = 0.03s
  get_tasks.py = 0.12s

Overall Massive Improvments!