feat: add scheduled task that updates task transitions in toggle block
This change adds a new function that is scheduled using APScheduler to
piggy-back on the Flask application's process to do work (for Heroku
this avoids using a secondary dyno).

This new function will check all tasks that are not completed or
completed in the current day. For each task, I record a transition if a
state change was detected (from the last status).

I ideally would have used an in-line database, but it doesn't seem to be
supported via notion-py. The current approach of using TextBlocks looks
like it'll do the trick though.

The format of `Next Up, 2019-08-11 21:11:00` will allow a chunk of these
to be processed later easily if needed.

To avoid any TimeZone confusion, the times are just local (without
showing it in the data).

Adding the --preloaded on the gunicorn in hopes that it helps somehow
with more workers (if that was ever needed). To be honest, doesn't seem
like it would hurt to add it.