feat: optimizing the notion-py calls by monkey-patching out caching
Taken from my previous work in alfred-notion:
 - https://github.com/kevinjalbert/alfred-notion/commit/74fba1bf486f694d22cddd9d4983c19f0443b936

The cache for notion-py is used to verify that nothing has changes since
the last snapshot since it collected data. Loading and saving the cache
took up some time in the startup/teardown of a call. The larger the
cache the longer it took to load back in.

In my case, everything is happening really quickly as short script
executions. I decided to monkey-patch out the saving/loading of the
cache for massive speed improvments.