feat: inital packagable solution of notion-scripts (basic functionality)
All the bundled functionality is a culimnation of work that has gone
into both alfred-notion and notion-heroku. The idea is to have a single
place where the notion scripts that interact with the API and my
template can exist. This way the different projects can use the same
foundational work.

At the same time, I ensured that there was better naming conventions
being used, as well as classes (opposed to the hacked together versions
in the other projects).