feat: add api security using a server stored token, use POST requests
The existing solution was using a non-secure API (anyone could access it
and thus add tasks/notes, assuming they knew the endpoints). This change
adds in a secret API_TOKEN that is set on the server, and is requried to
be passed in on each request.

I decided also to change the endpoints to accept POST body requests (for
easier content later on).