feat: improve performance getting tasks with fast tag local lookups
With this change I was seeing about a 2s (3.632s -> 1.694s) improvement
when fetching all my tasks.

The approach is basically avoiding additional work looking up the title
of each tag when rendering the task. We have the ID of the tags on the
task (as they are a relationship), and we have the tag data locally in
a json file. We are able to do the manual lookup ourselves to get the
tag.

The majority of this work was handled in the utils file, but this could
probably move out into its own class/file at somepoint.

Some tradeoffs in complexity for performance. Given I use the :search
command a lot, I aimed for performance.