feat: ability to view/update the current day's lights via API/Alfred
This feature allows the current day's lights to be viewed and updated
via the API or via Alfred. The meat of the work is done in the shared
NotionApi class.

A new command (:lights) has been added in the Alfred workflow.
A new endpoint GET /current_lights.json has been added to the server
A new endpoint POST /update_light.json has been added to the server