feat: add ability to update a collection block with multiple properties
The block_update function now accommodates multiple properties in the
json body of the request. This makes it work for both normal text blocks
which use the 'title' property as well as collection blocks.

I made the block_update use an atomic_transactions so that everything
updates or nothing does.

I fixed a missing update on the collection_view_content method.